docker-compose seperated and fixed influxdb, fastAPI runs natively now
