--- Database/.env
@@ -1,6 +1,0 @@
-INFLUXDB_USERNAME=rar
-INFLUXDB_PASSWORD=qwertz1234
-INFLUXDB_TOKEN=_6i-0rzenVKTsLYWGBjZ5St7SYZBE_cCeHEk8AcPe0fD1zSNZK5FboxGes5D4NRCGohR5QOsqwkV9ZlBqSEHQQ==
-INFLUXDB_ORG=IPA
-INFLUXDB_BUCKET=rokit-db
-INFLUXDB_URL=http://localhost:8086
