--- FastAPI/.nicegui/storage_user_e741d3e8-fae0-4f39-9b2f-a303369c6442.json
@@ -1,1 +1,0 @@
-{"dark_mode": false}
