+++ database/.env
@@ -1,0 +1,4 @@
+INFLUXDB_TOKEN=By_xvrxxTbSVIK7wmR9BYLjfOS4w6Jlv9RlyVOX5vSFt4EC9UQYqI3_XWGdzhTS3K5w2eblfmam_bkSAot0P9A==
+INFLUXDB_ORG=IPA
+INFLUXDB_BUCKET=rokit-db
+INFLUXDB_URL=http://localhost:8086
