--- docker/Dockerfile
@@ -1,15 +1,0 @@
-FROM python:3-slim-buster
-
-RUN mkdir /src
-WORKDIR /src
-
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
-
-CMD [ "uvicorn", "FastAPI.main:app --reload"]
-
-# COPY ./entrypoint.sh /
-# RUN chmod +x /entrypoint.sh
-# ENTRYPOINT ["/entrypoint.sh"]
