+++ docker/Dockerfile-nicegui
@@ -1,0 +1,11 @@
+FROM python:3-slim-buster
+
+RUN mkdir /src
+WORKDIR /src
+
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+CMD [ "uvicorn", "view.rokitAPI.main:app --reload", "--host", "localhost", "--port", "8000" ]
