+++ docker/docker-compose-influxdb.yaml
@@ -1,0 +1,10 @@
+version: '3.8'
+services:
+  influxdb:
+    image: influxdb:2.0.7
+    container_name: influxdb
+    ports:
+      - "8086:8086"
+    volumes:
+      - ./influxdb2:/var/lib/influxdb2
+      - ./config.yml:/etc/influxdb2/config.yml
