--- docker/docker-compose.yaml
@@ -1,27 +1,0 @@
-version: '3.8'
-services:
-  influxdb:
-    image: influxdb:2.0.7
-    container_name: influxdb
-    ports:
-      - "8086:8086"
-    volumes:
-    - ./influxdb2:/var/lib/influxdb2
-    - ./config.yml:/etc/influxdb2/config.yml
-    depends_on:
-    - nicegui
-    # command: 
-    #   influx setup \
-    #   --bucket rokit-db \
-    #   --org IPA \
-    #   --password qwertz1234 \
-    #   --username rar \
-    #   --force
-
-  nicegui:
-    build:
-      context: ..
-      dockerfile: docker/Dockerfile
-    container_name: nicegui
-    ports: 
-      - "8000:8000"
