--- docker/entrypoint.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-#echo "hello world"
-uvicorn FastAPI.main:app --reload
