--- docker/influxdb2/engine/data/6a9d6c574433bdbd/autogen/2/index/1/MANIFEST
@@ -1,37 +1,0 @@
-{
-  "levels": [
-    {},
-    {
-      "m": 33554432,
-      "k": 6
-    },
-    {
-      "m": 33554432,
-      "k": 6
-    },
-    {
-      "m": 67108864,
-      "k": 6
-    },
-    {
-      "m": 134217728,
-      "k": 6
-    },
-    {
-      "m": 268435456,
-      "k": 6
-    },
-    {
-      "m": 536870912,
-      "k": 6
-    },
-    {
-      "m": 1073741824,
-      "k": 6
-    }
-  ],
-  "files": [
-    "L0-00000001.tsl"
-  ],
-  "version": 1
-}
